+++ cppwamp/include/cppwamp/internal/callee.hpp
@@ -9,7 +9,6 @@
 
 #include <cstdint>
 #include <memory>
-#include <string>
 #include "../anyhandler.hpp"
 #include "../erroror.hpp"
 #include "../peerdata.hpp"
@@ -37,9 +36,13 @@ public:
         const Registration&,
         AnyCompletionHandler<void(ErrorOr<bool>)>&& handler) = 0;
 
-    virtual void safeYield(RequestId, wamp::Result&&) = 0;
+    virtual ErrorOrDone yield(RequestId, wamp::Result&&) = 0;
 
-    virtual void safeYield(RequestId, wamp::Error&&) = 0;
+    virtual std::future<ErrorOrDone> safeYield(RequestId, wamp::Result&&) = 0;
+
+    virtual ErrorOrDone yield(RequestId, wamp::Error&&) = 0;
+
+    virtual std::future<ErrorOrDone> safeYield(RequestId, wamp::Error&&) = 0;
 };
 
 } // namespace internal